Move Gateway validation func to make pkg structure symmetrical in shape
